Reference

Contents

Reference

Contents

Index

Index

RegularizedOptimization.ALSolver

RegularizedOptimization.AL

RegularizedOptimization.FISTA

RegularizedOptimization.LM

RegularizedOptimization.LMTR

RegularizedOptimization.PG

RegularizedOptimization.R2

RegularizedOptimization.R2DH

RegularizedOptimization.R2DH

RegularizedOptimization.R2N

RegularizedOptimization.RegularizedExecutionStats

RegularizedOptimization.TR

RegularizedOptimization.TRDH

RegularizedOptimization.project_y!

RegularizedOptimization.prox_split_1w

RegularizedOptimization.prox_split_2w

RegularizedOptimization.ALSolver — Type

AL(reg_nlp; kwargs...)

An augmented Lagrangian method for constrained regularized optimization, namely problems in the form

minimize    f(x) + h(x)
subject to  lvar ≤ x ≤ uvar,
            lcon ≤ c(x) ≤ ucon

where f: ℝⁿ → ℝ, c: ℝⁿ → ℝᵐ and their derivatives are Lipschitz continuous and h: ℝⁿ → ℝ is lower semi-continuous, proper and prox-bounded.

At each iteration, an iterate x is computed as an approximate solution of the subproblem

minimize    L(x;y,μ) + h(x)
subject to  lvar ≤ x ≤ uvar

where y is an estimate of the Lagrange multiplier vector for the constraints lcon ≤ c(x) ≤ ucon, μ is the penalty parameter and L(⋅;y,μ) is the augmented Lagrangian function defined by

L(x;y,μ) := f(x) - yᵀc(x) + ½ μ ‖c(x)‖².

For advanced usage, first define a solver "ALSolver" to preallocate the memory used in the algorithm, and then call solve!:

solver = ALSolver(reg_nlp)
solve!(solver, reg_nlp)

stats = GenericExecutionStats(reg_nlp.model)
solver = ALSolver(reg_nlp)
solve!(solver, reg_nlp, stats)

Arguments

reg_nlp::AbstractRegularizedNLPModel: a regularized optimization problem, see RegularizedProblems.jl, consisting of model representing a smooth optimization problem, see NLPModels.jl, and a regularizer h such as those defined in ProximalOperators.jl.

The objective and gradient of model will be accessed. The Hessian of model may be accessed or not, depending on the subsolver adopted. If adopted, the Hessian is accessed as an abstract operator and need not be the exact Hessian.

Keyword arguments

x::AbstractVector: a primal initial guess (default: reg_nlp.model.meta.x0)

y::AbstractVector: a dual initial guess (default: reg_nlp.model.meta.y0)

atol::T = √eps(T): absolute optimality tolerance;

ctol::T = atol: absolute feasibility tolerance;

verbose::Int = 0: if > 0, display iteration details every verbose iteration;

max_iter::Int = 10000: maximum number of iterations;

max_time::Float64 = 30.0: maximum time limit in seconds;

max_eval::Int = -1: maximum number of evaluation of the objective function (negative number means unlimited);

subsolver::AbstractOptimizationSolver = has_bounds(nlp) ? TR : R2: the procedure used to compute a step (e.g. PG, R2, TR or TRDH);

subsolver_logger::AbstractLogger: a logger to pass to the subproblem solver;

init_penalty::T = T(10): initial penalty parameter;

factor_penalty_up::T = T(2): multiplicative factor to increase the penalty parameter;

factor_primal_linear_improvement::T = T(3/4): fraction to declare sufficient improvement of feasibility;

init_subtol::T = T(0.1): initial subproblem tolerance;

factor_decrease_subtol::T = T(1/4): multiplicative factor to decrease the subproblem tolerance;

dual_safeguard = (nlp::AugLagModel) -> nothing: in-place function to modify, as needed, the dual estimate.

Output

stats::GenericExecutionStats: solution and other info, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in reg_nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.y: current dual estimate;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm has attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention;

stats.elapsed_time: elapsed time in seconds;

stats.solver_specific[:smooth_obj]: current value of the smooth part of the objective function;

stats.solver_specific[:nonsmooth_obj]: current value of the nonsmooth part of the objective function.

RegularizedOptimization.AL — Method

AL(reg_nlp; kwargs...)

An augmented Lagrangian method for constrained regularized optimization, namely problems in the form

minimize    f(x) + h(x)
subject to  lvar ≤ x ≤ uvar,
            lcon ≤ c(x) ≤ ucon

where f: ℝⁿ → ℝ, c: ℝⁿ → ℝᵐ and their derivatives are Lipschitz continuous and h: ℝⁿ → ℝ is lower semi-continuous, proper and prox-bounded.

At each iteration, an iterate x is computed as an approximate solution of the subproblem

minimize    L(x;y,μ) + h(x)
subject to  lvar ≤ x ≤ uvar

where y is an estimate of the Lagrange multiplier vector for the constraints lcon ≤ c(x) ≤ ucon, μ is the penalty parameter and L(⋅;y,μ) is the augmented Lagrangian function defined by

L(x;y,μ) := f(x) - yᵀc(x) + ½ μ ‖c(x)‖².

For advanced usage, first define a solver "ALSolver" to preallocate the memory used in the algorithm, and then call solve!:

solver = ALSolver(reg_nlp)
solve!(solver, reg_nlp)

stats = GenericExecutionStats(reg_nlp.model)
solver = ALSolver(reg_nlp)
solve!(solver, reg_nlp, stats)

Arguments

reg_nlp::AbstractRegularizedNLPModel: a regularized optimization problem, see RegularizedProblems.jl, consisting of model representing a smooth optimization problem, see NLPModels.jl, and a regularizer h such as those defined in ProximalOperators.jl.

The objective and gradient of model will be accessed. The Hessian of model may be accessed or not, depending on the subsolver adopted. If adopted, the Hessian is accessed as an abstract operator and need not be the exact Hessian.

Keyword arguments

x::AbstractVector: a primal initial guess (default: reg_nlp.model.meta.x0)

y::AbstractVector: a dual initial guess (default: reg_nlp.model.meta.y0)

atol::T = √eps(T): absolute optimality tolerance;

ctol::T = atol: absolute feasibility tolerance;

verbose::Int = 0: if > 0, display iteration details every verbose iteration;

max_iter::Int = 10000: maximum number of iterations;

max_time::Float64 = 30.0: maximum time limit in seconds;

max_eval::Int = -1: maximum number of evaluation of the objective function (negative number means unlimited);

subsolver::AbstractOptimizationSolver = has_bounds(nlp) ? TR : R2: the procedure used to compute a step (e.g. PG, R2, TR or TRDH);

subsolver_logger::AbstractLogger: a logger to pass to the subproblem solver;

init_penalty::T = T(10): initial penalty parameter;

factor_penalty_up::T = T(2): multiplicative factor to increase the penalty parameter;

factor_primal_linear_improvement::T = T(3/4): fraction to declare sufficient improvement of feasibility;

init_subtol::T = T(0.1): initial subproblem tolerance;

factor_decrease_subtol::T = T(1/4): multiplicative factor to decrease the subproblem tolerance;

dual_safeguard = (nlp::AugLagModel) -> nothing: in-place function to modify, as needed, the dual estimate.

Output

stats::GenericExecutionStats: solution and other info, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in reg_nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.y: current dual estimate;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm has attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention;

stats.elapsed_time: elapsed time in seconds;

stats.solver_specific[:smooth_obj]: current value of the smooth part of the objective function;

stats.solver_specific[:nonsmooth_obj]: current value of the nonsmooth part of the objective function.

RegularizedOptimization.FISTA — Method

FISTA for min_x ϕ(x) = f(x) + g(x), with f(x) cvx and β-smooth, g(x) closed cvx

Input: f: function handle that returns f(x) and ∇f(x) h: function handle that returns g(x) s: initial point proxG: function handle that calculates prox_{νg} options: see descentopts.jl Output: s⁺: s update s : s^(k-1) his : function history feval : number of function evals (total objective)

RegularizedOptimization.LM — Method

LM(nls, h, options; kwargs...)

A Levenberg-Marquardt method for the problem

min ½ ‖F(x)‖² + h(x)

where F: ℝⁿ → ℝᵐ and its Jacobian J are Lipschitz continuous and h: ℝⁿ → ℝ is lower semi-continuous, proper and prox-bounded.

At each iteration, a step s is computed as an approximate solution of

min  ½ ‖J(x) s + F(x)‖² + ½ σ ‖s‖² + ψ(s; x)

where F(x) and J(x) are the residual and its Jacobian at x, respectively, ψ(s; x) = h(x + s), and σ > 0 is a regularization parameter.

Arguments

nls::AbstractNLSModel: a smooth nonlinear least-squares problem

h: a regularizer such as those defined in ProximalOperators

options::ROSolverOptions: a structure containing algorithmic parameters

Keyword arguments

x0::AbstractVector: an initial guess (default: nls.meta.x0)

subsolver_logger::AbstractLogger: a logger to pass to the subproblem solver

subsolver: the procedure used to compute a step (PG, R2 or TRDH)

subsolver_options::ROSolverOptions: default options to pass to the subsolver.

selected::AbstractVector{<:Integer}: (default 1:nls.meta.nvar).

Return values

xk: the final iterate

Fobj_hist: an array with the history of values of the smooth objective

Hobj_hist: an array with the history of values of the nonsmooth objective

Complex_hist: an array with the history of number of inner iterations.

RegularizedOptimization.LMTR — Method

LMTR(nls, h, χ, options; kwargs...)

A trust-region Levenberg-Marquardt method for the problem

min ½ ‖F(x)‖² + h(x)

where F: ℝⁿ → ℝᵐ and its Jacobian J are Lipschitz continuous and h: ℝⁿ → ℝ is lower semi-continuous and proper.

At each iteration, a step s is computed as an approximate solution of

min  ½ ‖J(x) s + F(x)‖₂² + ψ(s; x)  subject to  ‖s‖ ≤ Δ

where F(x) and J(x) are the residual and its Jacobian at x, respectively, ψ(s; x) = h(x + s), ‖⋅‖ is a user-defined norm and Δ > 0 is a trust-region radius.

Arguments

nls::AbstractNLSModel: a smooth nonlinear least-squares problem

h: a regularizer such as those defined in ProximalOperators

χ: a norm used to define the trust region in the form of a regularizer

options::ROSolverOptions: a structure containing algorithmic parameters

Keyword arguments

x0::AbstractVector: an initial guess (default: nls.meta.x0)

subsolver_logger::AbstractLogger: a logger to pass to the subproblem solver

subsolver: the procedure used to compute a step (PG, R2 or TRDH)

subsolver_options::ROSolverOptions: default options to pass to the subsolver.

selected::AbstractVector{<:Integer}: (default 1:nls.meta.nvar).

Return values

xk: the final iterate

Fobj_hist: an array with the history of values of the smooth objective

Hobj_hist: an array with the history of values of the nonsmooth objective

Complex_hist: an array with the history of number of inner iterations.

RegularizedOptimization.PG — Method

Proximal Gradient Descent for

min_x ϕ(x) = f(x) + g(x), with f(x) β-smooth, g(x) closed, lsc

Input: f: function handle that returns f(x) and ∇f(x) h: function handle that returns g(x) s: initial point proxG: function handle that calculates prox_{νg} options: see descentopts.jl Output: s⁺: s update s : s^(k-1) his : function history feval : number of function evals (total objective )

RegularizedOptimization.R2 — Method

R2(reg_nlp; kwargs…)

A first-order quadratic regularization method for the problem

min f(x) + h(x)

where f: ℝⁿ → ℝ has a Lipschitz-continuous gradient, and h: ℝⁿ → ℝ is lower semi-continuous, proper and prox-bounded.

About each iterate xₖ, a step sₖ is computed as a solution of

min  φ(s; xₖ) + ½ σₖ ‖s‖² + ψ(s; xₖ)

where φ(s ; xₖ) = f(xₖ) + ∇f(xₖ)ᵀs is the Taylor linear approximation of f about xₖ, ψ(s; xₖ) is either h(xₖ + s) or an approximation of h(xₖ + s), ‖⋅‖ is a user-defined norm and σₖ > 0 is the regularization parameter.

For advanced usage, first define a solver "R2Solver" to preallocate the memory used in the algorithm, and then call solve!:

solver = R2Solver(reg_nlp)
solve!(solver, reg_nlp)

stats = RegularizedExecutionStats(reg_nlp)
solver = R2Solver(reg_nlp)
solve!(solver, reg_nlp, stats)

Arguments

reg_nlp::AbstractRegularizedNLPModel{T, V}: the problem to solve, see RegularizedProblems.jl, NLPModels.jl.

Keyword arguments

x::V = nlp.meta.x0: the initial guess;

atol::T = √eps(T): absolute tolerance;

rtol::T = √eps(T): relative tolerance;

neg_tol::T = eps(T)^(1 / 4): negative tolerance

max_eval::Int = -1: maximum number of evaluation of the objective function (negative number means unlimited);

max_time::Float64 = 30.0: maximum time limit in seconds;

max_iter::Int = 10000: maximum number of iterations;

verbose::Int = 0: if > 0, display iteration details every verbose iteration;

σmin::T = eps(T): minimum value of the regularization parameter;

η1::T = √√eps(T): very successful iteration threshold;

η2::T = T(0.9): successful iteration threshold;

ν::T = eps(T)^(1 / 5): multiplicative inverse of the regularization parameter: ν = 1/σ;

γ::T = T(3): regularization parameter multiplier, σ := σ/γ when the iteration is very successful and σ := σγ when the iteration is unsuccessful.

The algorithm stops either when √(ξₖ/νₖ) < atol + rtol*√(ξ₀/ν₀) or ξₖ < 0 and √(-ξₖ/νₖ) < neg_tol where ξₖ := f(xₖ) + h(xₖ) - φ(sₖ; xₖ) - ψ(sₖ; xₖ), and √(ξₖ/νₖ) is a stationarity measure.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.xk: current iterate;

solver.∇fk: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.solver_specific[:smooth_obj]: current value of the smooth part of the objective function

stats.solver_specific[:nonsmooth_obj]: current value of the nonsmooth part of the objective function

stats.status: current status of the algorithm. Should be :unknown unless the algorithm has attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

RegularizedOptimization.R2DH — Method

R2DH(f, ∇f!, h, options, x0)

A second calling form for R2DH where the objective and gradient are passed as arguments.

Arguments

f::Function: the objective function

∇f!::Function: the gradient function

h: a regularizer such as those defined in ProximalOperators

D: Diagonal quasi-Newton operator.

options::ROSolverOptions: a structure containing algorithmic parameters

x0::AbstractVector: an initial guess

Keyword Arguments

Mmonotone::Int: number of previous values of the objective to consider for the non-monotone variant (default: 6).

selected::AbstractVector{<:Integer}: subset of variables to which h is applied (default 1:length(x0)).

Return values

xk: the final iterate

k: the number of iterations

outdict: a dictionary containing the following fields:

Fhist: an array with the history of values of the smooth objective

Hhist: an array with the history of values of the nonsmooth objective

Time_hist: an array with the history of elapsed times

Chist: an array with the history of number of inner iterations

NonSmooth: the nonsmooth term

status: the status of the solver either :first_order, :max_iter, :max_time or :exception

fk: the value of the smooth objective at the final iterate

hk: the value of the nonsmooth objective at the final iterate

sqrt_ξ_νInv: the square root of the ratio of the nonsmooth term to the regularization parameter

elapsed_time: the elapsed time

RegularizedOptimization.R2DH — Method

R2DH(nlp, h, options)

A second-order quadratic regularization method for the problem

min f(x) + h(x)

where f: ℝⁿ → ℝ is C¹ and h: ℝⁿ → ℝ is lower semi-continuous, proper, and prox-bounded.

About each iterate xₖ, a step sₖ is computed as a solution of

min  φ(s; xₖ) + ψ(s; xₖ)

where φ(s ; xₖ) = f(xₖ) + ∇f(xₖ)ᵀs + ½ sᵀ(Dₖ + σₖI)s is a quadratic approximation of f about xₖ, ψ(s; xₖ) = h(xₖ + s), Dₖ is a diagonal Hessian approximation and σₖ > 0 is the regularization parameter.

Arguments

nlp::AbstractDiagonalQNModel: a smooth optimization problem

h: a regularizer such as those defined in ProximalOperators

options::ROSolverOptions: a structure containing algorithmic parameters

Keyword Arguments

x0::AbstractVector: an initial guess (in the first calling form: default = nlp.meta.x0)

selected::AbstractVector{<:Integer}: subset of variables to which h is applied (default 1:length(x0)).

D: Diagonal quasi-Newton operator.

Mmonotone::Int: number of previous values of the objective to consider for the non-monotone variant (default: 6).

The objective and gradient of nlp will be accessed.

Return values

The value returned is a GenericExecutionStats, see SolverCore.jl.

RegularizedOptimization.R2N — Method

R2N(nlp, h, χ, options; kwargs...)

A regularized quasi-Newton method for the problem

min f(x) + h(x)

where f: ℝⁿ → ℝ is C¹ and h: ℝⁿ → ℝ is lower semi-continuous, proper and prox-bounded.

About each iterate xₖ, a step sₖ is computed as an approximate solution of

min  φ(s; xₖ) + ½ σₖ ‖s‖² + ψ(s; xₖ)

where φ(s; xₖ) = f(xₖ) + ∇f(xₖ)ᵀs + ½ sᵀBₖs is a quadratic approximation of f about xₖ, ψ(s; xₖ) = h(xₖ + s) and σₖ > 0 is the regularization parameter. The subproblem is solved inexactly by way of a first-order method such as the proximal-gradient method or the quadratic regularization method.

Arguments

nlp::AbstractNLPModel: a smooth optimization problem

h: a regularizer such as those defined in ProximalOperators

options::ROSolverOptions: a structure containing algorithmic parameters.

The objective, gradient and Hessian of nlp will be accessed. The Hessian is accessed as an abstract operator and need not be the exact Hessian.

Keyword arguments

x0::AbstractVector: an initial guess (default: nlp.meta.x0)

subsolver_logger::AbstractLogger: a logger to pass to the subproblem solver (default: the null logger)

subsolver: the procedure used to compute a step (R2DH, R2 or PG)

subsolver_options::ROSolverOptions: default options to pass to the subsolver (default: all defaut options)

Mmonotone::Int: number of previous values of the objective to consider for the non-monotone variant (default: 1).

selected::AbstractVector{<:Integer}: subset of variables to which h is applied (default 1:nlp.meta.nvar).

Return values

xk: the final iterate

Fobj_hist: an array with the history of values of the smooth objective

Hobj_hist: an array with the history of values of the nonsmooth objective

Complex_hist: an array with the history of number of inner iterations.

RegularizedOptimization.RegularizedExecutionStats — Method

GenericExecutionStats(reg_nlp :: AbstractRegularizedNLPModel{T, V})

Construct a GenericExecutionStats object from an AbstractRegularizedNLPModel. More specifically, construct a GenericExecutionStats on the NLPModel of regnlp and add three solverspecific entries namely :smoothobj, :nonsmoothobj and :xi. This is useful for reducing the number of allocations when calling solve!(..., regnlp, stats) and should be used by default. Warning: This should not be used when adding other solverspecific entries that do not have the current scalar type.

RegularizedOptimization.TR — Method

TR(nlp, h, χ, options; kwargs...)

A trust-region method for the problem

min f(x) + h(x)

where f: ℝⁿ → ℝ has a Lipschitz-continuous Jacobian, and h: ℝⁿ → ℝ is lower semi-continuous and proper.

About each iterate xₖ, a step sₖ is computed as an approximate solution of

min  φ(s; xₖ) + ψ(s; xₖ)  subject to  ‖s‖ ≤ Δₖ

where φ(s ; xₖ) = f(xₖ) + ∇f(xₖ)ᵀs + ½ sᵀ Bₖ s is a quadratic approximation of f about xₖ, ψ(s; xₖ) = h(xₖ + s), ‖⋅‖ is a user-defined norm and Δₖ > 0 is the trust-region radius. The subproblem is solved inexactly by way of a first-order method such as the proximal-gradient method or the quadratic regularization method.

Arguments

nlp::AbstractNLPModel: a smooth optimization problem

h: a regularizer such as those defined in ProximalOperators

χ: a norm used to define the trust region in the form of a regularizer

options::ROSolverOptions: a structure containing algorithmic parameters

The objective, gradient and Hessian of nlp will be accessed. The Hessian is accessed as an abstract operator and need not be the exact Hessian.

Keyword arguments

x0::AbstractVector: an initial guess (default: nlp.meta.x0)

subsolver_logger::AbstractLogger: a logger to pass to the subproblem solver (default: the null logger)

subsolver: the procedure used to compute a step (PG, R2 or TRDH)

subsolver_options::ROSolverOptions: default options to pass to the subsolver (default: all defaut options)

selected::AbstractVector{<:Integer}: (default 1:f.meta.nvar).

Return values

xk: the final iterate

Fobj_hist: an array with the history of values of the smooth objective

Hobj_hist: an array with the history of values of the nonsmooth objective

Complex_hist: an array with the history of number of inner iterations.

RegularizedOptimization.TRDH — Method

TRDH(nlp, h, χ, options; kwargs...)
TRDH(f, ∇f!, h, options, x0)

A trust-region method with diagonal Hessian approximation for the problem

min f(x) + h(x)

where f: ℝⁿ → ℝ has a Lipschitz-continuous Jacobian, and h: ℝⁿ → ℝ is lower semi-continuous and proper.

About each iterate xₖ, a step sₖ is computed as an approximate solution of

min  φ(s; xₖ) + ψ(s; xₖ)  subject to  ‖s‖ ≤ Δₖ

where φ(s ; xₖ) = f(xₖ) + ∇f(xₖ)ᵀs + ½ sᵀ Dₖ s is a quadratic approximation of f about xₖ, ψ(s; xₖ) = h(xₖ + s), ‖⋅‖ is a user-defined norm, Dₖ is a diagonal Hessian approximation and Δₖ > 0 is the trust-region radius.

Arguments

nlp::AbstractDiagonalQNModel: a smooth optimization problem

h: a regularizer such as those defined in ProximalOperators

χ: a norm used to define the trust region in the form of a regularizer

options::ROSolverOptions: a structure containing algorithmic parameters

The objective and gradient of nlp will be accessed.

In the second form, instead of nlp, the user may pass in

f a function such that f(x) returns the value of f at x

∇f! a function to evaluate the gradient in place, i.e., such that ∇f!(g, x) store ∇f(x) in g

x0::AbstractVector: an initial guess.

Keyword arguments

x0::AbstractVector: an initial guess (default: nlp.meta.x0)

selected::AbstractVector{<:Integer}: (default 1:f.meta.nvar)

Return values

xk: the final iterate

Fobj_hist: an array with the history of values of the smooth objective

Hobj_hist: an array with the history of values of the nonsmooth objective

Complex_hist: an array with the history of number of inner iterations.

RegularizedOptimization.project_y! — Method

project_y!(nlp)

Given an AugLagModel, project nlp.y into [ymin, ymax] and updates nlp.μc_y accordingly.

RegularizedOptimization.prox_split_1w — Method

Solves descent direction s for some objective function with the structure mins qk(s) + ψ(x+s) s.t. ||s||q⩽ Δ for some Δ provided Arguments ––––– proxp : prox method for p-norm takes in z (vector), a (λ||⋅||p), p is norm for ψ I think s0 : Vector{Float64,1} Initial guess for the descent direction projq : generic that projects onto ||⋅||q⩽Δ norm ball options : mutable structure pparams

Returns

s : Vector{Float64,1} Final value of Algorithm 6.1 descent direction w : Vector{Float64,1} relaxation variable of Algorithm 6.1 descent direction

RegularizedOptimization.prox_split_2w — Method

Solves descent direction s for some objective function with the structure mins qk(s) + ψ(x+s) s.t. ||s||q⩽ Δ for some Δ provided Arguments ––––– proxp : prox method for p-norm takes in z (vector), a (λ||⋅||p), p is norm for ψ I think s0 : Vector{Float64,1} Initial guess for the descent direction projq : generic that projects onto ||⋅||q⩽Δ norm ball options : mutable structure pparams

Returns

s : Vector{Float64,1} Final value of Algorithm 6.2 descent direction w : Vector{Float64,1} relaxation variable of Algorithm 6.2 descent direction